+++ pymongo/asynchronous/command_cursor.py
@@ -382,6 +382,12 @@ class AsyncCommandCursor(Generic[_DocumentType]):
         await self.close()
 
     async def to_list(self) -> list[_DocumentType]:
+        """Converts the contents of this cursor to a list.
+
+        If the cursor is empty or has no more results, an empty list will be returned.
+
+        .. versionadded:: 4.9
+        """
         res: list[_DocumentType] = []
         while self.alive:
             if not await self._next_batch(res):